Moved cogs to utility.py
