+++ cogs/utility.py
@@ -7,6 +7,7 @@ import requests
 import json
 import discord
 import os
+import glob
 from PIL import Image
 from PythonGists import PythonGists
 from discord.ext import commands
@@ -614,20 +615,20 @@ class Utility:
         else:
             embed = discord.Embed(title="Number of people playing {}".format(game), description=msg)
             await self.bot.send_message(ctx.message.channel, "", embed=embed)
-            
+
     @commands.command(pass_context=True)
     async def animate(self, ctx, animation):
         """Play an animation from a text file. >help animate for more details.
         >animate <animation> - Animate a text file.
         Animation text files are stored in the anims folder. Each frame of animation is put on a new line.
-        
+
         An example text file looks like this:
         family
         fam ily
         fam i ly
         fam i love y
         fam i love you
-        
+
         You can additionally add a number to the top of the file to denote the delay between each frame. The default is 0.2 seconds.
         1
         fam
@@ -650,7 +651,7 @@ class Utility:
                 for frame in anim:
                     await asyncio.sleep(0.2)
                     await self.bot.edit_message(ctx.message, frame)
-                    
+
     @commands.command(pass_context=True)
     async def roles(self, ctx, *, user):
         """Check the roles of a user."""
@@ -670,7 +671,7 @@ class Utility:
             await self.bot.send_message(ctx.message.channel, "", embed=embed)
         else:
             await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "That user has no roles!")
-            
+
     @commands.command(pass_context=True)
     async def messagedump(self, ctx, limit, filename, details="yes", reverse="no"):
         """Dump messages."""
@@ -696,13 +697,13 @@ class Utility:
                     async for message in self.bot.logs_from(ctx.message.channel, int(limit), reverse=True):
                         f.write(message.content.encode() + "\n".encode())
         await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "Finished downloading!")
-        
+
     @commands.group(pass_context=True)
     async def link(self, ctx):
         await self.bot.delete_message(ctx.message)
         if ctx.invoked_subcommand is None:
             await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "Usage: `link <shorten/lengthen> <url>`")
-            
+
     @link.command(pass_context=True)
     async def shorten(self, ctx, url):
         try:
@@ -721,7 +722,7 @@ class Utility:
                 await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "There was an error shortening your URL.")
         else:
             await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "You did not enter a valid URL.")
-                
+
     @link.command(pass_context=True)
     async def lengthen(self, ctx, url):
         try:
@@ -732,7 +733,7 @@ class Utility:
             await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "<{}>".format(requests.get(url).url))
         else:
             await self.bot.send_message(ctx.message.channel, self.bot.bot_prefix + "You did not enter a valid URL.")
-            
+
     @commands.command(pass_context=True, aliases=['getcolor'])
     async def getcolour(self, ctx, colour_code):
         await self.bot.delete_message(ctx.message)
@@ -742,6 +743,39 @@ class Utility:
         image.save("colour_file.png")
         await self.bot.send_file(ctx.message.channel, "colour_file.png", content="Colour with hex code {}:".format(colour_code))
         os.remove("colour_file.png")
+
+    @commands.command()
+    async def cogs(self):
+        """Shows loaded/unloaded cogs"""
+
+        cog_list = ["cogs." + os.path.splitext(f)[0] for f in [os.path.basename(f) for f in glob.glob("cogs/*.py")]]
+
+        loaded = [x.__module__.split(".")[1] for x in self.bot.cogs.values()]
+        unloaded = [c.split(".")[1] for c in cog_list
+                    if c.split(".")[1] not in loaded]
+
+        loaded_formatted = []
+        for x in loaded:
+            loaded_formatted.append('+ {}'.format(x))
+
+        unloaded_formatted = []
+        if unloaded:
+            for x in unloaded:
+                unloaded_formatted.append('- {}'.format(x))
+        else:
+            unloaded_formatted = ['']
+
+        text = in_text = "```diff\nCogs:\n" + "\n".join(loaded_formatted) + "\n\n" + "\n".join(unloaded_formatted) + "```"
+        if len(in_text) > 2000:
+            while len(in_text) > 2000:
+                closest_delim = max([in_text.rfind(d, 0, 2000)
+                                     for d in delims])
+                closest_delim = closest_delim if closest_delim != -1 else 2000
+                to_send = in_text[:closest_delim]
+                await self.bot.say(to_send)
+                in_text = in_text[closest_delim:]
+        else:
+            await self.bot.say(text)
 
 def setup(bot):
     bot.add_cog(Utility(bot))
